+++ src/main/java/org/p2p/solanaj/rpc/types/RecentBlockhash.java
@@ -13,7 +13,7 @@ public class RecentBlockhash extends RpcResultObject {
     public static class FeeCalculator {
 
         @Json(name = "lamportsPerSignature")
-        private long lamportsPerSignature;
+        private double lamportsPerSignature;
     }
 
     @Getter
